requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
minio==7.2.16
pyyaml==6.0.2
jsonschema==4.25.1
